Implement native special item behavior
Add native-validated record reset, club member-capacity expansion, and Personal Board behavior with transactional inventory consumption and lifecycle tests.

Amp-Thread-ID: https://ampcode.com/threads/T-019ff605-8482-73d9-8f77-efbb235e2945

+++ chat-server/src/main/java/com/jftse/emulator/server/core/packets/item/S2CPersonalBoardMessageListPacket.java
@@ -6,12 +6,12 @@ import com.jftse.server.core.protocol.PacketOperations;
 import java.util.Map;
 
 public class S2CPersonalBoardMessageListPacket extends Packet {
-    public S2CPersonalBoardMessageListPacket(Map<Integer, String> messages) {
+    public S2CPersonalBoardMessageListPacket(Map<Short, String> messages) {
         super(PacketOperations.S2CPersonalBoardMessageList);
 
         this.write((short) messages.size());
-        for (Map.Entry<Integer, String> entry : messages.entrySet()) {
-            this.write(entry.getKey().shortValue());
+        for (Map.Entry<Short, String> entry : messages.entrySet()) {
+            this.write(entry.getKey());
             this.write(entry.getValue());
         }
     }
